JS-4175: space methods moved to UtilSpace
