+++ extension/iframe/create.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import $ from 'jquery';
 import { observer } from 'mobx-react';
 import { Button, Block, Loader, Icon, Select, IconObject, EmptySearch } from 'Component';
-import { I, C, M, translate, UtilObject, UtilData } from 'Lib';
+import { I, C, M, translate, UtilObject, UtilData, UtilSpace } from 'Lib';
 import { blockStore, extensionStore, menuStore, dbStore, commonStore } from 'Store';
 
 interface State {
@@ -103,7 +103,7 @@ const Create = observer(class Create extends React.Component<I.PageComponent, St
 
 	init () {
 		const spaces = dbStore.getSpaces()
-			.filter(it => it && UtilObject.canParticipantWrite(it.targetSpaceId))
+			.filter(it => it && UtilSpace. canParticipantWrite(it.targetSpaceId))
 			.map(it => ({ ...it, id: it.targetSpaceId, object: it, iconSize: 16 }));
 
 		if (this.refSpace && spaces.length) {
