+++ src/ts/component/page/main/edit.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { Header, Footer, EditorPage } from 'Component';
-import { I, Onboarding, UtilObject, analytics } from 'Lib';
+import { I, Onboarding, UtilObject, analytics, UtilSpace } from 'Lib';
 import { detailStore, blockStore } from 'Store';
 
 class PageMainEdit extends React.Component<I.PageComponent> {
@@ -34,7 +34,7 @@ class PageMainEdit extends React.Component<I.PageComponent> {
 	onOpen () {
 		const { isPopup } = this.props;
 		const rootId = this.getRootId();
-		const home = UtilObject.getSpaceDashboard();
+		const home = UtilSpace.getDashboard();
 		const object = detailStore.get(rootId, rootId, [ 'type' ], true);
 
 		if (this.refHeader) {
