+++ src/ts/component/page/main/media.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import $ from 'jquery';
 import { observer } from 'mobx-react';
 import { Header, Footer, Loader, Block, Button, IconObject, Deleted } from 'Component';
-import { I, C, UtilCommon, Action, Renderer, UtilObject, translate, UtilRouter } from 'Lib';
+import { I, C, UtilCommon, Action, Renderer, UtilSpace, translate, UtilRouter } from 'Lib';
 import { blockStore, detailStore } from 'Store';
 import HeadSimple from 'Component/page/elements/head/simple';
 import Errors from 'json/error.json';
@@ -187,7 +187,7 @@ const PageMainMedia = observer(class PageMainMedia extends React.Component<I.Pag
 				if (message.error.code == Errors.Code.NOT_FOUND) {
 					this.setState({ isDeleted: true });
 				} else {
-					UtilObject.openHome('route');
+					UtilSpace.openDashboard('route');
 				};
 				return;
 			};
