+++ src/ts/component/popup/search.tsx
@@ -3,7 +3,7 @@ import $ from 'jquery';
 import { observer } from 'mobx-react';
 import { AutoSizer, CellMeasurer, InfiniteLoader, List, CellMeasurerCache } from 'react-virtualized';
 import { Icon, Loader, IconObject, ObjectName, EmptySearch, Label, Filter } from 'Component';
-import { I, UtilCommon, UtilData, UtilObject, UtilRouter, keyboard, Key, focus, translate, analytics, Action } from 'Lib';
+import { I, UtilCommon, UtilData, UtilObject, UtilRouter, keyboard, Key, focus, translate, analytics, Action, UtilSpace } from 'Lib';
 import { commonStore, dbStore, popupStore } from 'Store';
 import Constant from 'json/constant.json';
 
@@ -58,10 +58,10 @@ const PopupSearch = observer(class PopupSearch extends React.Component<I.Popup, 
 				object = item;
 			} else 
 			if (item.id == 'account') {
-				object = UtilObject.getParticipant();
+				object = UtilSpace.getParticipant();
 			} else 
 			if (item.id == 'spaceIndex') {
-				object = UtilObject.getSpaceview();
+				object = UtilSpace.getSpaceview();
 			};
 
 			if (object) {
@@ -413,7 +413,7 @@ const PopupSearch = observer(class PopupSearch extends React.Component<I.Popup, 
 		const hasRelations = keyboard.isMainEditor() || keyboard.isMainSet();
 		const filter = this.getFilter();
 		const lang = Constant.default.interfaceLang;
-		const canWrite = UtilObject.canParticipantWrite();
+		const canWrite = UtilSpace. canParticipantWrite();
 
 		let name = '';
 		if (filter) {
